move GasLimitCalculator interface to ethereum, add targetGasLimit to MiningParameters
