+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/GasLimitCalculator.java
@@ -12,7 +12,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.ethereum.blockcreation;
+package org.hyperledger.besu.ethereum;
 
 public interface GasLimitCalculator {
